Solution to edge case
Solves the edge case where a variable from an enclosing scope is used to initialize a local with the same name

+++ compiler.h
@@ -11,7 +11,7 @@
 
 struct Local {
 	Token token;
-	size_t depth;
+	int depth;
 };
 
 // Considering splitting this into a parser and compiler
